Moved wallets to market
